revised job scripts to check for output and abort if output already exists so qc_framework doesn't start over everytime

+++ job_scripts/run_aligner.sh
@@ -18,15 +18,22 @@ echo no OUTPUT set! stop
 exit 1
 fi
 
-
-echo aligning file $INPUT
-/slipstream/galaxy/production/galaxy-dist/tools/star/STAR --genomeLoad NoSharedMemory --genomeDir /slipstream/galaxy/data/hg38/star_index --readFilesIn $INPUT --runThreadN 8 --alignIntronMax 1 --outSAMtype BAM SortedByCoordinate --outStd BAM_SortedByCoordinate > $OUTPUT
-mv Log.final.out $OUTPUT".log"
+if [ -f $OUTPUT ]; then
+	echo $OUTPUT exists, skipping alignment for $INPUT
+else
+	echo aligning file $INPUT
+	/slipstream/galaxy/production/galaxy-dist/tools/star/STAR --genomeLoad NoSharedMemory --genomeDir /slipstream/galaxy/data/hg38/star_index --readFilesIn $INPUT --runThreadN 8 --alignIntronMax 1 --outSAMtype BAM SortedByCoordinate --outStd BAM_SortedByCoordinate > $OUTPUT
+	mv Log.final.out $OUTPUT".log"
+fi
 #echo i am an alignment of $INPUT > $OUTPUT
 #echo trimming file $INPUT
 
 echo aligned file is written to $OUTPUT
-echo indexing file $OUTPUT
-samtools index $OUTPUT
+if [ -f $OUTPUT".bai" ]; then
+	echo $OUTPUT".bai" exists, skipping indexing for $OUTPUT
+else
+	echo indexing file $OUTPUT
+	samtools index $OUTPUT
+fi
 ls -lha $OUTPUT*
 echo done
